moves specs_substitutions
