+++ packages/models-library/src/models_library/utils/specs_substitution.py
@@ -1,0 +1,66 @@
+from typing import Any
+
+import yaml
+from models_library.utils.string_substitution import (
+    SubstitutionsDict,
+    TextTemplate,
+    substitute_all_legacy_identifiers,
+)
+
+from .string_substitution import SubstitutionsDict, TextTemplate
+
+# This constraint is to avoid deserialization issues after substitution!
+SubstitutionValues = bool | int | float | str
+
+
+class SpecsEnvironmentsResolver:
+    def __init__(self, specs: dict[str, Any], upgrade: bool):
+        self._template = self._create_text_template(specs, upgrade=upgrade)
+        self._substitutions: SubstitutionsDict = SubstitutionsDict()
+
+    @classmethod
+    def _create_text_template(
+        cls, specs: dict[str, Any], *, upgrade: bool
+    ) -> TextTemplate:
+        # convert to yaml (less symbols as in json)
+        service_spec_str: str = yaml.safe_dump(specs)
+
+        if upgrade:  # legacy
+            service_spec_str = substitute_all_legacy_identifiers(service_spec_str)
+
+        # template
+        template = TextTemplate(service_spec_str)
+        assert template.is_valid()  # nosec
+
+        return template
+
+    def get_identifiers(self) -> list[str]:
+        """lists identifiers in specs in order of apperance. Can have repetitions"""
+        return self._template.get_identifiers()
+
+    def get_replaced(self) -> set[str]:
+        return self._substitutions.used
+
+    @property
+    def substitutions(self):
+        return self._substitutions
+
+    def set_substitutions(
+        self, environs: dict[str, SubstitutionValues]
+    ) -> SubstitutionsDict:
+        identifiers_needed = self.get_identifiers()
+
+        # picks only needed for substitution
+        self._substitutions = SubstitutionsDict(
+            {
+                identifier: environs[identifier]
+                for identifier in identifiers_needed
+                if identifier in environs
+            }
+        )
+        return self._substitutions
+
+    def run(self) -> dict[str, Any]:
+        new_specs_txt: str = self._template.safe_substitute(self._substitutions)
+        new_specs: dict = yaml.safe_load(new_specs_txt)
+        return new_specs
